Finalized New Normalization of Account Data
The account data lacked so much information, I had to first compare to existing DB data, then search the API for anything missing. Now the functions to normalize account data and normalize extended history return data in the same format. Up next is probably finding a way to load this to the DB.